refactor: utility types
