+++ packages/shared/src/components/notifications/utils.ts
@@ -1,4 +1,11 @@
+import { ComponentType } from 'react';
 import classed from '../../lib/classed';
+import { IconProps } from '../Icon';
+import BellIcon from '../icons/Bell';
+import CommunityPicksIcon from '../icons/CommunityPicksIcon';
+import DailyIcon from '../icons/DailyIcon';
+import DiscussIcon from '../icons/Discuss';
+import UpvoteIcon from '../icons/Upvote';
 
 export const NotifContainer = classed(
   'div',
@@ -16,3 +23,30 @@ export const NotifProgress = classed(
   'span',
   'absolute -bottom-2 h-1 ease-in-out bg-theme-status-cabbage rounded-full',
 );
+
+export enum NotificationIcon {
+  DailyDev = 'DailyDev',
+  CommunityPicks = 'CommunityPicks',
+  Comment = 'Comment',
+  Upvote = 'Upvote',
+  Bell = 'Bell',
+}
+
+export const notificationIcon: Record<
+  NotificationIcon,
+  ComponentType<IconProps>
+> = {
+  [NotificationIcon.DailyDev]: DailyIcon,
+  [NotificationIcon.CommunityPicks]: CommunityPicksIcon,
+  [NotificationIcon.Comment]: DiscussIcon,
+  [NotificationIcon.Upvote]: UpvoteIcon,
+  [NotificationIcon.Bell]: BellIcon,
+};
+
+export const notificationDefaultTheme: Record<NotificationIcon, string> = {
+  [NotificationIcon.DailyDev]: '',
+  [NotificationIcon.CommunityPicks]: '',
+  [NotificationIcon.Comment]: 'text-theme-color-blueCheese',
+  [NotificationIcon.Upvote]: 'text-theme-color-blueCheese',
+  [NotificationIcon.Bell]: '',
+};
